requirements.txt
requests>=2.25.0
pycryptodome>=3.15.0
beautifulsoup4>=4.9.0
urllib3>=1.26.0
Pillow>=8.0.0
